+++ components/Header/Header.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux'
 import Link from 'next/link'
+import Router from 'next/router'
 import { Button, Dropdown, Menu } from 'semantic-ui-react'
 
 // import { Input } from 'semantic-ui-react'
@@ -17,18 +18,22 @@ class Header extends React.Component {
   
   handleItemClick = (e, { name }) => this.setState({ activeItem: name })
 
+  handleLogin = (e) => Router.push('http://127.0.0.1/oauth/authorize/?client_id=2&redirect_uri=http://127.0.0.1:3000/token&response_type=code&scope=access-user-account')
+
   render() {
     const { profile } = this.props;
-    const { auth, activeItem } = this.state;
+    const { activeItem } = this.state;
+    // console.log(profile);
 
-    const userIcon = <img src="./static/assets/images/Icons/icon-user.svg" />
-    const loginLink = 'http://127.0.0.1/oauth/authorize/?client_id=2&redirect_uri=http://127.0.0.1:3000/token&response_type=code&scope=access-user-account';
+    const userIcon = <img src="http://localhost:3000/static/assets/images/Icons/icon-user.svg" />
 
     return (
       <Menu color="red" fixed="top" inverted className="SiteHeader">
-        <Menu.Item name='logo' active={activeItem === 'logo'} onClick={this.handleItemClick}>
-          <i className="icon logo" alt="Kushy logo"></i>
-        </Menu.Item>
+        <Link href="/" passHref>
+          <Menu.Item as="a" name='logo' active={activeItem === 'logo'}>
+            <i className="icon logo" alt="Kushy logo"></i>
+          </Menu.Item>
+        </Link>
         <Menu.Item
           name='search'
         >
@@ -41,23 +46,21 @@ class Header extends React.Component {
           {profile ? 
             <Dropdown item trigger={userIcon}>
               <Dropdown.Menu>
-                <Link href={ '/user/' + profile.username }>
-                <Dropdown.Item>
-                    Dashboard
-                </Dropdown.Item>
+                <Link href={ '/dashboard/' } passHref>
+                  <Dropdown.Item>
+                      Dashboard
+                  </Dropdown.Item>
                 </Link>
-                  <Link href={ '/logout' }>
-                <Dropdown.Item>
-                    Logout
-                </Dropdown.Item>
+                  <Link href={ '/logout' } passHref>
+                    <Dropdown.Item>
+                        Logout
+                    </Dropdown.Item>
                   </Link>
               </Dropdown.Menu>
             </Dropdown>
           : 
             <Menu.Item>
-              <Link href={loginLink}>
-                <Button primary>Sign Up</Button>
-              </Link>
+                <Button primary onClick={ this.handleLogin }>Login</Button>
             </Menu.Item>}
         </Menu.Menu>
       </Menu>