+++ config/config.js
@@ -1,4 +1,5 @@
 export default {
     tokenName: 'kushyFToken',
-    kushyApiUrl: 'http://localhost/api/v1'
+    kushyApiUrl: 'http://localhost/api/v1',
+    kushyLogin: 'http://127.0.0.1/oauth/authorize/?client_id=2&redirect_uri=http://127.0.0.1:3000/token&response_type=code&scope=access-user-account',
 }