+++ pages/_app.js
@@ -1,7 +1,7 @@
 import App, {Container} from 'next/app'
 import React from 'react'
 import { PersistGate } from 'redux-persist/lib/integration/react';
-import withReduxStore from '../lib/with-redux-store'
+import withReduxStore from 'lib/with-redux-store'
 import { Provider } from 'react-redux'
 
 class MyApp extends App {